Correct Kore configuration.
Kore ran on a single worker in previous benchmarks.

This is due Kore by default accepting as many new connections
as it can. Considering the max connections was set to 250 and
the benchmark was run with 100 connections only this resulted
in only a single worker doing all the heavy lifting.

The new settings will limit workers their accept loop to
a maximum of 10 new connections at a time before giving
up on the accept lock so others can grab connections too.

Also kills setting affinity as the author decided to use
12 worker processes.